feat(zoo): redesign model registry as card grid + per-model detail page
- /zoo/registry now shows a tiny-card grid of model groups (two rows
  per card: name + version-count + latest-tag chips on row 1,
  description on row 2; edit/delete icons on hover).
- /zoo/registry/:groupId is a new detail page with a header
  (back button, title, description, metadata pills, edit/delete) and
  Model Card / Versions tabs.
- Versions table dropped the redundant "Name" column; Model ID moved
  to a tooltip on the Version cell.
- Top-bar alignment fixed (no more &nbsp; FormLabel hack).
- Splits the old ModelRegistry.tsx into ModelRegistryGrid,
  ModelGroupCard, ModelGroupVersionsTable, ModelGroupDetail.
